优化 和 数据库字段类型的修改

+++ jpress-service/src/main/java/io/jpress/service/AttachmentVideoCategoryService.java
@@ -178,5 +178,5 @@ public interface AttachmentVideoCategoryService  {
      * 根据category_id 更新model
      * @param categoryId
      */
-    void doUpdateVideoCategoryCount(Integer categoryId);
+    void doUpdateVideoCategoryCount(Long categoryId);
 }
